Set testimony_get_block timeout to 1000ms
With no timeout, the while loop causes high strain on the CPU by flooding it with poll syscalls. I set it to 1000ms since Clerk and Stenographer use the same function and set the timeout to 1000ms. In my testing, the timeout fixed CPU usage and had no noticeable affect elsewhere.